+++ src/Symfony/Component/Messenger/TraceableMessageBus.php
@@ -29,30 +29,22 @@ class TraceableMessageBus implements MessageBusInterface
      */
     public function dispatch($message): void
     {
-        $caller = $this->getCaller();
-        $callTime = microtime(true);
-        $messageToTrace = $message instanceof Envelope ? $message->getMessage() : $message;
-        $envelopeItems = $message instanceof Envelope ? array_values($message->all()) : null;
+        $envelope = Envelope::wrap($message);
+        $context = array(
+            'envelopeItems' => array_values($envelope->all()),
+            'message' => $envelope->getMessage(),
+            'caller' => $this->getCaller(),
+            'callTime' => microtime(true),
+        );
 
         try {
             $this->decoratedBus->dispatch($message);
-
-            $this->dispatchedMessages[] = array(
-                'envelopeItems' => $envelopeItems,
-                'message' => $messageToTrace,
-                'callTime' => $callTime,
-                'caller' => $caller,
-            );
         } catch (\Throwable $e) {
-            $this->dispatchedMessages[] = array(
-                'envelopeItems' => $envelopeItems,
-                'message' => $messageToTrace,
-                'exception' => $e,
-                'callTime' => $callTime,
-                'caller' => $caller,
-            );
+            $context['exception'] = $e;
 
             throw $e;
+        } finally {
+            $this->dispatchedMessages[] = $context;
         }
     }
 
@@ -82,10 +74,7 @@ class TraceableMessageBus implements MessageBusInterface
                 $line = $trace[$i]['line'];
 
                 while (++$i < 8) {
-                    if (isset($trace[$i]['function'], $trace[$i]['file']) && empty($trace[$i]['class']) && 0 !== strpos(
-                            $trace[$i]['function'],
-                            'call_user_func'
-                        )) {
+                    if (isset($trace[$i]['function'], $trace[$i]['file']) && empty($trace[$i]['class']) && 0 !== strpos($trace[$i]['function'], 'call_user_func')) {
                         $file = $trace[$i]['file'];
                         $line = $trace[$i]['line'];